pyproject.toml
[project]
name = "currencybot"
version = "1.0.0"
description = "Currency bot for KiwiNest"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  "python-dotenv==1.1.1",
  "discord-py==2.6.3",
  "aiosqlite==0.21.0",
]

[tool.ruff]
line-length = 130
indent-width = 4

[tool.ruff.lint]
# Require explicit selection of preview rules.
explicit-preview-rules = true
select = ["ALL"]
extend-select = ["ALL"]
extend-safe-fixes = ["ALL"]
extend-ignore = [
  "C901",
  "T201",   # Ignore print
  "DTZ005",
  "N806",   # allow dunder
  "S311",   # allow non-crypto random
  "INP001", # implicit namespace
  "D100",   # Missing Docstring
  "D101",   # Missing Docstring
  "D102",   # Missing Docstring
  "D103",   # Missing Docstring
  "D107",   # Missing Docstring
  "FBT001", # boolean-default-value-positional-argument
  "FBT002", # boolean-default-value-positional-argument
  "FBT003", # boolean-default-value-positional-argument
  "ERA001", # Commented out code
]
preview = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
